MainProject_v1/projectSettings/App/file_test.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework import status
from django.http import JsonResponse

@api_view(["POST"])
def test(sentence):
    print(sentence)